Restore the block engine to main and make the fluid sim the default
The block model, its DOM renderer and its tests go back to their state on
main, dropping the body-settling water rewrite and the three.js renderer of
the block world. The FLIP fluid sim keeps its three.js renderer and is now
the front page, with the block game at /dom and a toggle in the app bar
replacing the navigation drawer.

The restored liquid tests were failing on main itself: addActiveBlock only
queues a block, so each test now opens with one process call to flush the
queue before expecting water to move.
